feat(mcp): advertise output schemas for core read tools
Adopt mcp-go's structured-output support for roady's core read tools:
roady_get_plan, roady_get_state, roady_get_spec, and roady_get_snapshot now
declare an OutputSchema, so tools/list advertises their output shape and the
server emits structuredContent. MCP clients get typed, validated project data
(plan, execution state, product spec, snapshot) instead of parsing JSON text.

These handlers already returned their typed domain values via (any, error), so
this is almost purely additive — the only code change is hoisting the
handler-local snapshotResp type to package scope (output_schemas.go) so it can
be advertised.

A guard test asserts each advertised output type generates a schema: OutputSchema
runs schema.Generate at registration, and a failure there would silently drop
the tool (the ToolBuilder short-circuits), which the direct-call handler tests
would not catch.

Follows the same house convention as mnemos/briefkasten/chronos (return the
typed struct + OutputSchema). Template for the remaining read tools
(roady_get_usage, roady_forecast, roady_get_ready_tasks, ...). Recursive domain
types generate cleanly on v1.21; the v1.22 bump (#55 follow-up) further improves
them via JSON Schema 2020-12 $ref/$defs.
